refactor: extract rewind buffer to minarch_rewind.c/h
Move the rewind buffer system (~800 lines) out of minarch.c into
dedicated minarch_rewind.c/h files. Public API functions lose their
static qualifier; private helpers retain it. The REWIND_STEP_* enum
and MINARCH_DEFAULT_REWIND_* / REWIND_* defines are now in the header.
minarch_internal.h includes minarch_rewind.h so all modules see the
RewindContext type and extern declaration.
