+++ workspace/all/minarch/minarch_internal.h
@@ -98,3 +98,5 @@ extern int rewind_cfg_compress;
 extern int rewind_cfg_lz4_acceleration;
 
 bool getAlias(char* path, char* alias);
+
+#include "minarch_rewind.h"
